Download third party libraries for travis from private repo.

+++ ThirdPartyLibs/ASL/wgetASL.sh
@@ -4,16 +4,20 @@ echo " "
 echo "##### Downloading the third party packages for PIPS-NLP:"
 echo " "
 
-echo "### Downloading Ampl Solver Library (ASL):"
-if wget -O solvers.tar.gz http://www.ampl.com/netlib/ampl/solvers.tgz
-then
-  echo "### ASL: Download Successful.\n"
+fn=solvers.tar.gz
+if [ ! -f $fn ]; then
+  echo "### Downloading Ampl Solver Library (ASL):"
+  if wget -O solvers.tar.gz http://www.ampl.com/netlib/ampl/solvers.tgz
+  then
+    echo "### ASL: Download Successful.\n"
+  else
+    echo "### ASL: Download Failed.\n"
+    exit 1 
+  fi
 else
-  echo "### ASL: Download Failed.\n"
-  exit 1 
+  echo "File found. No download."
 fi
 
-fn=solvers.tar.gz
 name=`basename ${fn} .tar.gz`
 tar -xzf $fn
 ln -s ./${name} ./src
